refactor tu support multiple brands
