--- src/caenhv/__init__.py
@@ -1,1 +1,0 @@
-from .devices import CaenHV
