--- src/caenhv/devices/__init__.py
@@ -1,1 +1,0 @@
-from .caenhv import CaenHV
