+++ src/hvps/commands/caen/__init__.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 import re
 
-from .module import _get_set_module_command, _get_mon_module_command
+from src.hvps.commands.caen.module import _get_set_module_command, _get_mon_module_command
 
 import logging
