+++ src/hvps/commands/caen/module.py
@@ -25,7 +25,7 @@ def _get_mon_module_command(bd: int, parameter: str) -> bytes:
     Generate a command string to monitor a specific module parameter.
 
     Args:
-        bd (int): The board number.
+        bd (int): The board number. Must be in the range 0..31.
         parameter (str): The parameter to monitor.
 
     Returns:
@@ -34,6 +34,9 @@ def _get_mon_module_command(bd: int, parameter: str) -> bytes:
     Raises:
         ValueError: If the provided parameter is not valid.
     """
+    if not 0 <= bd <= 31:
+        raise ValueError(f"Invalid board number '{bd}'. Must be in the range 0..31.")
+
     parameter = parameter.upper()
     if parameter not in _mon_module_parameters:
         valid_parameters = ", ".join(_mon_module_parameters.keys())
