+++ src/hvps/devices/channel.py
@@ -1,9 +1,7 @@
-from functools import cached_property
-
 import serial
 
-from ..commands.channel import _get_set_channel_command, _get_mon_channel_command
-from ..commands import _parse_response
+from src.hvps.commands.caen.channel import _get_set_channel_command, _get_mon_channel_command
+from src.hvps.commands.caen import _parse_response
 
 from time import sleep
